fold spawn_tuple each into spawn_tuple
